// src/app/Models/Cliente.ts
export  interface Cliente{
    id:number;
    nombre: string;
    direccion: string;
    identificacion:string;
    telefono:string;
    correo: string;
    celular: string;
    sexo:  string;
    imagen:  string;
    message:string;
    isSuccess:boolean;
    result:string;
    clientes: Cliente[];
    
}